BirthdayCalendar renaming widget, creating manager class, turning widget to become stateless, using ChangeProvider
